refactor: [#264] split show command info DTOs into service submodules
- Move ServiceInfo to info/tracker.rs submodule
- Move PrometheusInfo to info/prometheus.rs submodule
- Move GrafanaInfo to info/grafana.rs submodule
- Keep EnvironmentInfo and InfrastructureInfo in info/mod.rs
- Add tests for PrometheusInfo and GrafanaInfo
- This structure makes it easier to add/remove services to the stack
